+++ Source/Core/Core/Core.vcxproj
@@ -81,16 +81,16 @@
     <ClCompile Include="DSP\Interpreter\DSPIntLoadStore.cpp" />
     <ClCompile Include="DSP\Interpreter\DSPIntMisc.cpp" />
     <ClCompile Include="DSP\Interpreter\DSPIntMultiplier.cpp" />
-    <ClCompile Include="DSP\Jit\DSPEmitter.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitArithmetic.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitBranch.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitCCUtil.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitExtOps.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitLoadStore.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitMisc.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitMultiplier.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitRegCache.cpp" />
-    <ClCompile Include="DSP\Jit\DSPJitUtil.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPEmitter.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitArithmetic.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitBranch.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitCCUtil.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitExtOps.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitLoadStore.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitMisc.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitMultiplier.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitRegCache.cpp" />
+    <ClCompile Include="DSP\Jit\x64\DSPJitUtil.cpp" />
     <ClCompile Include="DSP\LabelMap.cpp" />
     <ClCompile Include="FifoPlayer\FifoAnalyzer.cpp" />
     <ClCompile Include="FifoPlayer\FifoDataFile.cpp" />
@@ -339,8 +339,8 @@
     <ClInclude Include="DSP\Interpreter\DSPInterpreter.h" />
     <ClInclude Include="DSP\Interpreter\DSPIntExtOps.h" />
     <ClInclude Include="DSP\Interpreter\DSPIntUtil.h" />
-    <ClInclude Include="DSP\Jit\DSPEmitter.h" />
-    <ClInclude Include="DSP\Jit\DSPJitRegCache.h" />
+    <ClInclude Include="DSP\Jit\x64\DSPEmitter.h" />
+    <ClInclude Include="DSP\Jit\x64\DSPJitRegCache.h" />
     <ClInclude Include="DSP\LabelMap.h" />
     <ClInclude Include="FifoPlayer\FifoAnalyzer.h" />
     <ClInclude Include="FifoPlayer\FifoDataFile.h" />
