+++ Source/Core/Core/DSP/DSPTables.cpp
@@ -14,7 +14,7 @@
 
 #include "Core/DSP/Interpreter/DSPIntExtOps.h"
 #include "Core/DSP/Interpreter/DSPInterpreter.h"
-#include "Core/DSP/Jit/DSPEmitter.h"
+#include "Core/DSP/Jit/x64/DSPEmitter.h"
 
 namespace DSP
 {
