+++ Source/Core/Core/DSP/Jit/x64/DSPJitArithmetic.cpp
@@ -8,7 +8,7 @@
 
 #include "Core/DSP/DSPCore.h"
 #include "Core/DSP/DSPMemoryMap.h"
-#include "Core/DSP/Jit/DSPEmitter.h"
+#include "Core/DSP/Jit/x64/DSPEmitter.h"
 
 using namespace Gen;
