+++ Source/Core/Core/DSP/Jit/x64/DSPJitExtOps.cpp
@@ -5,7 +5,7 @@
 #include "Common/CommonTypes.h"
 
 #include "Core/DSP/DSPCore.h"
-#include "Core/DSP/Jit/DSPEmitter.h"
+#include "Core/DSP/Jit/x64/DSPEmitter.h"
 
 using namespace Gen;
