+++ Source/Core/Core/DSP/Jit/x64/DSPJitUtil.cpp
@@ -6,7 +6,7 @@
 
 #include "Core/DSP/DSPCore.h"
 #include "Core/DSP/DSPHWInterface.h"
-#include "Core/DSP/Jit/DSPEmitter.h"
+#include "Core/DSP/Jit/x64/DSPEmitter.h"
 
 using namespace Gen;
